+++ src/stack/avg_stack.c
@@ -6,7 +6,7 @@
 /*   By: unite <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/13 17:56:18 by unite             #+#    #+#             */
-/*   Updated: 2020/05/13 18:30:05 by unite            ###   ########.fr       */
+/*   Updated: 2020/05/15 03:27:41 by unite            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	avg_stack(const t_stack *stack)
 	link = stack->start;
 	while (link)
 	{
-		avg += link->value / stack->size;
+		avg += (long double)link->value / (long double)stack->size;
 		link = link->next;
 	}
 	return (avg);
